Update ntee.csv
Updates to language in descriptions for organizations that deal with "mental retardation". 

G25

Voluntary health organizations active in the prevention or treatment of Down Syndrome, a hereditary condition that results in mental retardation.

Voluntary health organizations active in the prevention or treatment of Down Syndrome, a hereditary condition that results in intellectual disability.

H25

Organizations that conduct research which can be used to improve the prevention, diagnosis and treatment of Down Syndrome, a hereditary condition that results in mental retardation.

Organizations that conduct research which can be used to improve the prevention, diagnosis and treatment of Down Syndrome, a hereditary condition that results in intellectual disability.

P82

Organizations such as retarded citizens associations that provide or coordinate a wide variety of programs and services for people with developmental disabilities.

Organizations that provide or coordinate a wide variety of programs and services for people with developmental disabilities.